pi-sensors/requirements.txt
requests==2.31.0
RPi.GPIO==0.7.1
Adafruit_DHT==1.4.0